end shop

+++ controllers/admin.js
@@ -1,7 +1,9 @@
 const mongodb = require('mongodb')
 const Product = require('../models/product');
-const { validationResult } = require('express-validator/check')
-const ObjectId = mongodb.ObjectId
+const { validationResult } = require('express-validator')
+const mongoose = require('mongoose')
+const fileHelper = require('../util/file');
+const product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
@@ -19,10 +21,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title
-  const imageUrl = req.body.imageUrl
+  const image = req.file
   const price = req.body.price
   const description = req.body.description
   const errors = validationResult(req)
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+    });
+  }
+
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -39,6 +58,9 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path
+
   const product = new Product({
     title: title,
     price: price,
@@ -53,7 +75,9 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      console.log(err);
+      const error = new Error(err)
+      error.httpStatusCode = 500
+      return next(error)
     })
 };
 
@@ -79,7 +103,9 @@ exports.getEditProduct = (req, res, next) => {
       });
     })
     .catch(err => {
-      console.log(err);
+      const error = new Error(err)
+      error.httpStatusCode = 500
+      return next(error)
     });
 };
 
@@ -87,7 +113,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updateTitle = req.body.title
   const updatePrice = req.body.price
-  const updateImageUrl = req.body.imageUrl
+  const image = req.file
   const updateDescription = req.body.description
   const errors = validationResult(req)
 
@@ -99,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updateTitle,
-        imageUrl: updateImageUrl,
         price: updatePrice,
         description: updateDescription,
         _id: prodId
@@ -117,14 +142,20 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updateTitle
       product.price = updatePrice
       product.description = updateDescription
-      product.imageUrl = updateImageUrl
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
       return product.save()
         .then(result => {
           res.redirect('/admin/products')
         })
     })
     .catch(err => {
-      console.log(err);
+      const error = new Error(err)
+      error.httpStatusCode = 500
+      return next(error)
     })
 }
 
@@ -137,19 +168,28 @@ exports.getProducts = (req, res, next) => {
         path: '/admin/products',
       });
     }).catch(err => {
-      console.log(err);
+      const error = new Error(err)
+      error.httpStatusCode = 500
+      return next(error)
     });
 }
 
 
 // // Delete product
-exports.postDeleteProduct = (req, res, next) => {
-  const productId = req.body.productId
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+  const productId = req.params.productId
+  Product.findById(productId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'))
+      }
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: productId, userId: req.user._id })
+    })
     .then(() => {
-      res.redirect('/admin/products')
+      res.status(200).json({ message: 'Success!' })
     })
     .catch(err => {
-      console.log(err);
+      res.status(500).json({ message: 'Deleting product failed' })
     })
 }
